+++ packages/jaspr/CHANGELOG.md
@@ -1,3 +1,15 @@
+## Unreleased patch
+
+- Improved the stability and logging of the cli and added the commands:
+  - `clean` command to clean your project directory
+  - `update` command to automatically update the cli to the latest version
+  - `doctor` command to print information about the environment and project
+  
+- We added lightweight anonymous usage tracking to the cli. We use [mixpanel.com](https://mixpanel.com/home) and
+  only process anonymized data. The usage statistics are made public and can be viewed here (TODO: Link will be added in next release).
+
+  To opt out of usage tracking, use `jaspr --disable-analytics`.
+
 ## 0.9.0
 
 - Added *Static Site Generation* support.